:pencil: Added readme and PR changes for validation section
